Split proofread inference out of auth module
Keep OAuth, session storage, and model prefs in auth.rs; move the system
prompt and provider Responses API calls into inference.rs so login and
proofreading stay separate while auth still supplies tokens.
